Correctly handle conversion of defaults from string to type
The previous approach bypassed the taking of the action. Now,
the overrides are stored as defaults, which pass through the argparse
machinery to parse strings. The downside is that parse_args now modifies
the parser, but parsers are generally single-shot anyway. I tried
copy.deepcopy to prevent this, but it refused to copy a parser.