Add timeout to catch connection errors
It is sometimes tricky to determine if a redis server is available.
This adds a Timeout class using Unix signals a la
http://stackoverflow.com/a/22348885/942503, which wraps the first
command to redis during construction of the TelescopeState object.
Without it, the subscribe command hangs in the absence of a server
(at least on OS X).

I tried hard to raise a redis.ConnectionError instead of a custom
TimeoutError but this somehow interfered with the signals.