Provide fake local telstate if endpoint is empty
If the endpoint is an empty string or omitted (the default), build
TelescopeState on top of a local instance of redis provided by the
fakeredis package. This supports all of the relevant redis commands
and stores the key-value pairs in local memory, so unit tests will
still pass. It also serves as a useful fallback option for script
dry-running and SDP simulators that need the telstate API without
talking to an actual server.