Allow wait condition to access timestamps
The current wait_key() condition only checks the value of a telstate
item. Extend it to include the associated timestamp of the item
(which defaults to the current time for attributes without timestamps).
This allows scripts to wait on telstate until a fresh value appears
for an item.

In the process, extend the redis pubsub update message format to include
the timestamp as a packed struct (very much like the raw sensor format).
This exposed an ASCII encoding bug in fakeredis on Python 2, for which
we include a monkey patch as workaround.

Tests pass on Python 2.7.13 and 3.6.0.