Allow wait condition to access timestamps
The current wait_key() condition only checks the value of a telstate
item. Extend it to optionally check the associated timestamp of the
item (if the key is mutable). This allows scripts to wait on telstate
until a fresh value appears for an item.

In the process, extend the redis pubsub update message format to match
the raw sensor format used in the main key-value store. In other words,
attributes have pickled values and sensors additionally prepend a packed
timestamp to each pickled value.

The timestamp parameter of the wait_key condition is optional so that
this extension is backwards compatible.

This exposed an ASCII encoding bug in fakeredis on Python 2, for which
we include a monkey patch as workaround. Once fakeredis issue #146
is resolved the monkey patch can go away.

This addresses JIRA ticket SR-965.