Allow wait_key to be manually cancelled
This allows a future object to be passed into wait_key, which when done
will cancel the wait and raise an exception. For now it is just polled,
but it is possible that in future it could do something intelligent to
get a callback as soon as the future is set.