Add endpoints_to_str function
This is the inverse of endpoints_list_parser, and turns a list of
endpoints into a compact addr+N:port string.

This code was originally part of katcbfsim, but with split ingest coming
it will be useful in a number of other places e.g. for more compact
logging and passing around more compact command-line arguments.