Fix wait_key corner case with multiple namespaces
The _check_condition function was not properly namespace aware, and
would assume that the message data went with the value of
is_immutable(key), for example, even though they could involve different
prefixes; and it would take the message data as "the" value, even if it
was already shadowed by an earlier prefix.

The new version bypasses all the higher-level machinery and iterates
over the full keys on redis until it finds a match.

This won't immediately work in unit testing, because the unit test runs
into a fakenewsredis bug which will require a new release.

Closes SR-999