Introduce katsdptelstate.ConnectionError
During construction the TelecopeState object connects to the real Redis
server in order to subscribe to pub-sub messages. This will crash with
a redis error if the server is unavailable (or the endpoint is bad).

Instead of saddling the user with an explicit redis dependency, rather
wrap the redis exception in a custom katsdptelstate exception. The user
can then catch the new ConnectionError when telstate is unavailable.

This does hide the stack trace (i.e. consider `raise from` in future),
but the two exceptions involved tend to be self-explanatory. In the
case of TimeoutError the message is even improved by providing the
endpoint.