Improve handling of mixed unicode/bytes
Previously passing a `bytes` as a key in Python 3 could fail because
constructing the full key was done by adding a prefix and the key. This
occurred in katdal when it iterated over telstate.keys (which returns
bytes).

It now more closely follows the redis-py model that keys **are** bytes,
but can be **specified** as unicode. All prefixes and keys are coerced
to bytes (UTF-8 encoding) on entry, before adding them together.

The main user-visible change is that telstate.prefixes will now return
bytes-coerced versions of the prefixes given to the constructor. Some
exception messages may also be a bit less readable in Python 3 because
they'll say `b'key'` instead of `key`.