Provide helper functions for encoding and decoding
While the actual encoding remains the same (pickle), an abstraction is
provided by encode_value and decode_value functions, together with
EncodeError and DecodeError exceptions which should be used in place of
the raft of other raft of possible exceptions that cPickle can raise.

Once user code has been converted to use these wrappers it will be
possible to switch to something like msgpack for new data (see SR-1366).

This is not strictly backwards-compatible because the `return_pickle`
keyword arg and the `equal_pickles` function have been renamed to
reflect their more general function. However, from a quick search it
appears that only katsdpingest is affected.