Add alternative msgpack encoding for telstate
It is not the default, as it will require some time until all telstate
consumers switch to the generic decode_value.

Unlike the pickle encoding, it doesn't allow arbitrary code execution,
and is also resistant to denial-of-service via encoding an excessive
length. The fuzz test was made stronger (it had a bug where bytes were
only replaced with 255, and it's extended with truncation tests) which
turned up this weakness in pickle.

The encoding tests are also made stricter by requiring an exact
round-trip of the type. This is made possible by an extension encoding
for numpy scalars that encodes them as a dtype plus raw value rather
than letting them be converted to the msgpack core types.