Make compatible with fakeredis 1.0+
I'm shortly going to release fakeredis 1.0, which removes the singleton
option (and makes it default).

setup.py still specifies <1.0 because 1.0 is much slower (particularly a
problem when using it to access data in an RDB file, where it is at
least 5x slower and uses almost double the RAM). But this will at least
avoid crashes if someone forcibly installs a newer version.