Separate out a redis backend for telstate
This will make it possible to add other backends - particularly a
read-only backend that lazily loads data from a .rdb file, and hence is
more efficient when not all keys are read.

The send_message/get_message functions no longer take a channel. It
wouldn't have worked with a non-default channel anywhere, because pubsub
was never subscribed to it. Possibly we can kill the feature entirely.

There still needs to be some re-work on the main constructor so that one
can construct one's own backend.