Regenerate doc-requirements.txt
Use pip-compile -U to get the latest packages and versions, and then
tweak each version to match the one in katsdpdockerbase's
base-requirements.txt. This is mainly to get six to play nicely, since
the existing 1.11 pin interfered with the 1.12 setup.py requirement
in the Jenkins tests.

Also add typing requirement for Python 2 only.