Make MemoryBackend the default
There were some fairly substantial changes to the unit tests, mostly
because I noticed that the load_from_file tests weren't in great shape.
They're now split into a lower-level (redis layer) test of
tabloid_redis, rdb_reader, rdb_writer; and a higher-level test of
TelescopeState.load_from_file (per backend).

Also make fakeredis an optional dependency, since it's now only needed
for testing and for a few special cases (like constructing keys in
memory to write to an RDB file), and remove the pin on <1.0 (1.0 is
slower, but no longer on the critical path). For testing, require 1.0.1
to avoid the need to handle the `singleton` argument differently between
0.x and 1.0, and 1.0.1 fixes a bug that was crashing the test.