Make it possible to write to RDB from a MemoryBackend
The Backend API gains a `dump` function, which for redis is just a
wrapper around the redis-py `dump` function, but on MemoryBackend
constructs the dump. The TabloidRedis code mostly moves into rdb_utility
so that it can be shared by MemoryBackend. RDBWriter can now take a
Backend instead of a redis client (and in future that should probably be
preferred).

At this point TabloidRedis has very little value, and is preserved only
for backwards compatibility until katsdpmetawriter no longer depends on
it.