Fix encoding of large elements in zsets
The prev-length field was being encoded as 8-byte big-endian instead of
4-byte little-endian (the endianness isn't documented in the RDB dump
file format, but is described in the redis source).

Additionally, if there is >4GB of data in total then the ziplist
encoding can't be used.

Also changed the struct.pack specifiers to unsigned, to avoid risks of
other issues.