Load RDB files from file-like objects
This extends :meth:`TelescopeState.load_from_file` to load RDB files
from file-like objects (as well as from filenames). This supports
the ability to load RDB files directly from the Ceph store into memory.

Change the interface of :func:`rdb_reader.load_from_file` so that it
accepts a callback object instead of a client object. This lets the
memory and Redis backends share the same function. In the case of
a file object, get the size by seeking to the end of the file.

This supports JIRA ticket SR-1443.