Test that redis errors don't become RdbParseError
Use the ability of fakeredis to mock a connection error to verify that
these kinds of errors triggered in the RDB callback are not converted
into `RdbParseError`. To help with this, refactor RDB writing into
a helper method (`save_to_file`).

In the process, fix `TestLoadFromFileRedis` to actually contain tests
(it was subclassed from `TestCase` instead of `TestLoadFromFile`).