Introduce BackendCallback base class
Now that the Callback classes are essentially `RdbCallback` classes
with two custom attributes (`n_keys` and `client_busy`), make that a
new base class instead. It sure simplifies the convoluted explanation
of what to pass to `rdb_reader.load_from_file`. Instead of the bland
`Callback` and `_Callback` names, disambiguate them by including the
name of the backend for which they are customised, i.e. `RedisCallback`
and `MemoryCallback`.