Recast RDBWriter as a file resource
Instead of constructing an RDBWriter instance around a Redis client
and then passing the filename to the `save` method, turn it all around.
Construct the RDBWriter as a context manager around the RDB file itself
and pass the client to `save` instead. This removes the need for
separate "supplemental" routines to append extra telstate keys while
still terminating the file properly. The supplemental keys are now also
treated the same way as the usual keys (e.g. skipped if not found).

In the process, the `encode_item` function is moved outside the class
and simplified to be more like `encode_len`. This function has always
been an undeclared static method anyway. Also, instead of returning
the number of written and failed keys from `save`, rather tally them
as attributes on the RDBWriter object so that supplemental keys are
counted too.