Allow the separator to be changed
When establishing a TelescopeState, it attempts to write a key
\xffseparator (the prefix is chosen to avoid conflict with any possible
UTF-8-encoded string). If it already existed, the existing value is used
as the separator instead. The default choice of separator is not changed
now.

Once this version is used everywhere, the separator can be changed
without worrying about conflicts, because all users of a single
telescope state will negotiate a single separator (whichever arrived
first). The only time it will go wrong is when merging databases in
RDBWriter, where two databases may have different separators.

Partly implements SR-1832.