Add KeyUpdate class
The many tuple varieties were getting out of hand and making the code
hard to read. Now there is a class hierarchy so that fields can be
references by name.

I also spotted a bug in _check_condition (it was ignoring key updates
for mutable when it should have been ignoring them for indexed), and
extended the existing unit test to catch it.