Stop throwing InvalidKeyError
Since keys can be (and generally are) accessed by `__getitem__` rather
than `__getattr__`, it is not necessary to disallow using telstate keys
that happen to match method names. It's also problematic because adding
a new method breaks backwards compatibility as it makes a previously
valid key become invalid; and it will become more problematic if we add
additional clients (e.g. asyncio support) as the set of invalid keys
will depend on the interface used.

Remove the check from `add` and `__setitem__`. I've left a check in
`__setattr__` since if an attribute is set with `__setattr__` then it
will likely be retrieved again with `__getattr__` which won't work. The
check has also been improved to check for class members at all levels of
the object's class hierarchy, not just the most-derived class.

I've left the InvalidKeyError exception class in place to avoid breaking
code that catches it, but it will not be thrown.