Convert aio telstate to actually be asyncio
The frontend is sprinkled with async and await, and the memory backend
is rewritten to just wrap a synchronous backend. A quick test shows the
memory backend basically working, but there are no proper tests yet and
I'm not at all sure about wait_keys.

This now only works with Python 3.6 due to the use of an asynchronous
generator.