Multiplex aio pubsubs over pool's standard connection
The previous approach would consume a connection from the pool for every
concurrent use of monitor_keys; with the default of 10 that proved
insufficient.

Instead, use the built-in pub-sub connection provided by the pool for
all pub-sub commands. This is more complex, because
(a) multiple calls to monitor_keys may watch the same key (handled by
using a custom pubsub Channel with a set of queues to broadcast to).
(b) we need to unsubscribe from channels only once there are no further
users.

I think this should be safe, but there are some headache-inducing corner
cases due to the windows between issuing a (un)subscription and getting
the response. The channel only appears in
self.client.connection.pubsub_channels after aioredis has processed the
response. So for example that we'll see an existing channel and start
to use it, while at the same time it's being unsubscribed. Fortunately
aioredis notifies the channel when the unsubscription arrives, so it can
poison the queues and trigger monitor_keys to go around the loop again.

Closes SPR1-644.