Use a dedicated thread for async MemoryBackend.monitor_keys
The monitoring process takes an unbounded amount of time, so if run on
the default thread pool executor it can tie up all the available
threads.

This has the potential to create a larger-than-ideal number of threads,
but monitoring keys on a memory backend is only really useful for unit
tests.