Rewrite of aioredis support for aioredis 2.x
aioredis 2 is based on redis-py, so the code gets a little closer to the
redis-py backend. I left in the command re-execution, although it's less
critical since aioredis 2 now has the same health checks as redis-py. It
would be good to eventually port that to the redis-py backend.

I also rewrote the monitor_keys demultiplexer. While that could also be
ported to the redis-py backend, it's probably less needed there (since
there are unlikely to be large numbers of threads) and it would be more
complicated to write since pre-emptive concurrency is harder than
cooperative concurrency.

I also discovered a bug that we've somehow got away with so far: if an
indexed key is being monitored with `wait_key_indexed`, and a
*different* subkey is updated, it triggers an assertion.