Another rewrite of pub/sub handling for aioredis
This is more lines of code but easier to reason about. All interaction
with the pubsub object is funnelled through a single coroutine, which
should be safer (the aioredis code for dealing with the case where a new
connection needs to be established didn't look particularly async-safe).
monitor_keys no longer handles subscription or unsubscription directly;
it sends messages to that coroutine via a queue to request subscription
or unsubscription.