Work around an aioredis bug
If the pubsub connection is allowed to unsubscribe from everything, then
later (after health check interval expires) tries to resubscribe, it can
end up with a bogus message.

The workaround is belt-and-braces: always subscribe to some dummy
channel (with a name constructed to be unlikely to conflict with
any sane use), and also detect and ignore the bogus messages.