SPR1-2766 Upgrade to redis-py 4.2.0
This version now includes asyncio support, so there is no more need for
the separate aioredis package. The interfaces have remained the same as
in aioredis 2.x, so it is a drop-in replacement. It necessitates bumping
the minimum Python version to 3.6 (which is already dead anyway).

Upgrade fakeredis to 2.0.0+ in the process, which depends on redis>=4.2
and works with the integrated asyncio support instead of the aioredis
package. The tests will still need Python 3.7 (already true for the
previous release due to `pytest-asyncio >= 0.17.0`).

This upgrade is needed to fix some regressions in the aio unit tests
(12 failures in `aio/test/test_telescope_state.py`), involving the
following exceptions on immutable and indexed responses::

  - redis.exceptions.NoScriptError: No matching script. Please use EVAL.
  - redis.exceptions.ResponseError: WRONGTYPE Operation against a key
    holding the wrong kind of value
  - TypeError: 'ResponseError' object is not iterable

This seems to be an interaction between aioredis 2.0.1 and recent
redis-py versions.