SPR1-2807 Restrict versions of redis and fakeredis
Redis-py 4.2.0 absorbed aioredis and presents a breakpoint in the
redis Python universe. The corresponding version of fakeredis is 2.0.0.
By restricting ourselves to earlier versions, we get aioredis 2.x
support without rocking the boat yet. Leave that for v0.13.

In the process, determine that the minimum version of `pytest-asyncio`
is 0.17.0 due to the use of the `@pytest_asyncio.fixture` decorator.
This actually implies that the tests need Python 3.7...