SPR1-2766 Fix warnings in the unit tests
Change a nose-style method to the equivalent pytest method.

Ignore DeprecationWarnings and FutureWarnings that arise while
messing around with an encoded string at the byte level and trying
to break it. This is bound to lead to strange results, and if these
warnings ever become more serious they will hopefully be converted
to DecodeErrors.

Do version checks with packaging instead of distutils. This adds
`packaging` as a test dependency.