SPR1-3236 Fix tests for fakeredis >= 2.21.3
The docker base still uses fakeredis 2.10.0. When I bumped fakeredis to
the latest 2.29.0, I found that `fakeredis.FakeRedis()` doesn’t give
you a fresh Redis instance anymore but reuses one underneath the hood,
which messes up the katsdptelstate unit tests by persisting state.

I have to instantiate a new FakeServer to get a fresh telstate:
```
fakeredis.FakeRedis(server=fakeredis.FakeServer())
```

I traced this back to fakeredis 2.17.0 and cunla/fakeredis-py#213.
Because FakeRedis uses a connection pool, it makes one FakeServer with
a random name and then reuses it afterwards, unless you generate your
own FakeServer.

With this change (needed on sync side only), fakeredis 2.21.3+ works
for both sync and aio versions of telstate. I did not change the
requirement yet since `fakeredis < 2.12.0` is also fine. Some of the
versions in between also hang on the aio side. Once the docker base
bumps fakeredis we can require `fakeredis >= 2.21.3`.

The leaking of state did however uncover a buglet. :-) We intended to
test the assignment of an immutable value to a pre-existing indexed key.
Unfortunately that key was generated as a mutable instead. Because the
leaked state already had "test_indexed" as an indexed key from a prior
test, it complained when it inadvertently tried to become a mutable.
Yay for bugs uncovering other bugs!